Move QSfmlCanvas to qwidgets folder
